refactor: create /shared/skeletons directory and move SkeletonDiv
