+++ src/shared/ui/index.ts
@@ -9,3 +9,4 @@ export * from './select-categories'
 export * from './spacer'
 export * from './tab-button'
 export * from './back-button'
+export * from './skeletons'
