+++ src/shared/ui/skeletons/index.ts
@@ -1,0 +1,1 @@
+export * from './skeleton-div'
