+++ src/shared/ui/skeletons/skeleton-div.tsx
